Making the container div: Heirarchy:
Among the three positioning values (fixed, absolute, and relative), only fixed positions an element relative to the viewport.

Absolute positions an element relative to its closest positioned ancestor (other elements with absolute or relative or fixed positioning ), and

Relative positions an element relative to its normal position in the document flow.

The transition-all class in Tailwind CSS applies a transition effect to all CSS properties when they change. It essentially sets a transition for all properties that are changed during an interaction or state change.

The duration-500 class specifies the duration of the transition, in this case, 500 milliseconds. It determines how long the transition effect will take to complete. In this example, duration-500 sets the transition duration to 500 milliseconds, meaning any property changes will transition over half a second.

also added the cut icon from font awesome cdn