Prepared the logic: Since container's width is changed transsition and duration is applied . Width changes from 0 to 50vw which is 50% of view port
threeDots.classList.toggle('hidden') toggles the presence of the CSS class hidden on the threeDots element. If the hidden class is not present, it will be added; if it is already present, it will be removed.

In this context, it seems that the hidden class is used to control the visibility of the threeDots element. Toggling this class will show or hide the three dots based on its current visibility state. If the dots are currently hidden, this function call will make them visible, and vice versa.

.hidden {
    display: none;
}

The width property directly accessed from the element (container.width) isn't used to set the width of an element in the same way as style.width. The width property accessed this way returns the width of the element, not its style value, and it's read-only.

To dynamically change the width of an element using JavaScript, you typically use the style property to access the inline CSS styles of the element and modify them, as demonstrated with container.style.width. This allows you to directly control the rendering of the element on the page.