Prepared the layout with some key points in mind:
The reason why position: fixed; is necessary for this code to work is because left and top positioning properties only have an effect on elements that have their position property set to absolute, fixed, relative, or sticky. By setting position: fixed; on myDiv,  its positioning is calculated relative to the browser window(viewPort), rather than its containing element or the document itself.