myDiv.style.left = x - 100 + "px";:
This line sets the CSS left property of the myDiv element. In CSS, the left property is used to position an element horizontally relative to its containing block. Here, x - 100 calculates the new horizontal position of the element based on the X-coordinate of the mouse cursor (x) minus 100 pixels. The result is concatenated with "px" to convert it into a valid CSS value representing pixels.

myDiv.style.top = y - 100 + "px";: This line sets the CSS top property of the myDiv element. Similar to left, the top property is used to position an element vertically relative to its containing block. Here, y - 100 calculates the new vertical position of the element based on the Y-coordinate of the mouse cursor (y) minus 100 pixels. Again, the result is concatenated with "px" to convert it into a valid CSS value representing pixels.

In this case, subtracting 100 pixels means that the center of the myDiv element (which contains the rocket image) will be positioned 100 pixels to the left and 100 pixels above the mouse cursor. This offset creates a visually pleasing effect where the rocket appears to follow the mouse cursor with its center.